Combine fMRI/MEG into hebart2023things
